fix problem

+++ BTS project/Assignment 8/Assignment 8/Controllers/Product_vm.cs
@@ -21,7 +21,7 @@ namespace Assignment_8.Controllers
         public string ProductName { get; set; }
 
         [Display(Name = "Price")]
-        public decimal ProductPrice { get; set; }
+        public double ProductPrice { get; set; }
 
         [DataType(DataType.MultilineText)]
         [Display(Name = "Description")]
@@ -52,7 +52,7 @@ namespace Assignment_8.Controllers
         public int PromotionId { get; set; }
         public Promotion_vm Promotion { get; set; }
         [Display(Name ="Promotion Price")]
-        public decimal PromoPrice { get; set; }
+        public double PromoPrice { get; set; }
         public List<Promotion_vm> Promotions { get; set; }
     }
 }
